requirements.txt
Django==3.2
psycopg2==2.8.6
pytz==2021.1	
sqlparse==0.4.1
gunicorn==19.9.0
python-dotenv==0.17.1
django-storages==1.11.1
djangorestframework==3.12.4
django-colorfield==0.4.1
django-filter==2.4.0














